Move Saml2Error
Move to core package

Closes gh-8835
